chore: rename button component file
